+++ src/EventSubscriber/HomepageBrowserLanguageRedirection.php
@@ -97,14 +97,14 @@ class HomepageBrowserLanguageRedirection implements EventSubscriberInterface {
       return;
     }
 
-    // Whether or not preventing redirection when Referer Header is given.
+    // Whether preventing redirection when Referer Header is given.
     $referer_bypass_enabled = (bool) $this->configFactory->get('home_redirect_lang.browser_fallback')->get('enable_referer_bypass');
 
     /** @var \Drupal\Core\Language\Language $current_language */
     $current_language = $this->languageManager->getCurrentLanguage();
     $http_referer = $this->request->server->get('HTTP_REFERER');
     $current_host = $this->request->getHost();
-    $referer_host = parse_url($http_referer, \PHP_URL_HOST);
+    $referer_host = parse_url((string) $http_referer, \PHP_URL_HOST);
 
     // Ensure the REFERER is external to disable redirection.
     if ($referer_bypass_enabled && !empty($referer_host) && !empty($current_host) && $current_host !== $referer_host) {