+++ CHANGELOG.md
@@ -14,7 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - add coverage for Drupal 9.3, 9.4 & 9.5
-- drop support of drupal below 9.3.x
+- add official support of drupal 9.5 & 10.0
+
+### Changed
+- re-enable PHPUnit Symfony Deprecation notice
 
 ## [1.0.0-alpha1] - 2022-09-22
 ### Added